After much consideration, the game will not be multiplayer.
The reason for this is that I would like to have a game of a playable size that I actually 'release' or at least finish. Networking my previous projects, and this one, takes up almost triple the time that it takes to add those base features. Not only that, but debugging and testing becomes exponentially more difficult as the project grows. Adding features and real content, mechanics and play time would take me so long that the project would struggle to grow to the scale I want it to. For all these reasons, the game will never have multiplayer support.