Finally implementing the projectile angle fix.
Initially was going to do the Short Range Fix featured in NQD, but actually there is a much simpler and cleaner way to do things: just point the projectile at the mouse, from the muzzle, and clamp the projectile angle offset to a reasonable range compared to a straight line angle from the shooter's center of mass. Gives the same result, with the projectile passing right though the mouse position, but also prevents stupid shooting angles (such as being able to shoot backwards into the gun).